Movendo classes de Gif para um pacote próprio. Traduzindo as classes de gif para português e tornando os métodos e atributos públicos
