+++ src/br/univali/portugol/nucleo/bibliotecas/Graficos.java
@@ -6,9 +6,9 @@ import br.univali.portugol.nucleo.bibliotecas.base.ErroExecucaoBiblioteca;
 import br.univali.portugol.nucleo.bibliotecas.base.TipoBiblioteca;
 import br.univali.portugol.nucleo.bibliotecas.base.anotacoes.*;
 import br.univali.portugol.nucleo.bibliotecas.graficos.CacheImagens;
-import br.univali.portugol.nucleo.bibliotecas.graficos.ImageFrame;
 import br.univali.portugol.nucleo.bibliotecas.graficos.JanelaGrafica;
 import br.univali.portugol.nucleo.bibliotecas.graficos.JanelaGraficaImpl;
+import br.univali.portugol.nucleo.bibliotecas.graficos.gif.ImagemGif;
 import java.awt.*;
 import java.awt.event.FocusListener;
 import java.awt.event.KeyListener;
@@ -17,23 +17,15 @@ import java.awt.geom.AffineTransform;
 import java.awt.geom.Point2D;
 import java.awt.image.AffineTransformOp;
 import java.awt.image.BufferedImage;
-import java.awt.image.ColorModel;
 import java.awt.image.WritableRaster;
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.imageio.ImageIO;
-import javax.imageio.ImageReader;
-import javax.imageio.metadata.IIOMetadata;
-import javax.imageio.metadata.IIOMetadataNode;
-import org.w3c.dom.NamedNodeMap;
-import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
 
 /**
  *
@@ -82,7 +74,7 @@ public final class Graficos extends Biblioteca implements Teclado.InstaladorTecl
     private JanelaGrafica janela;
     private CacheImagens cacheImagens;
     private boolean inicializado = false;
-    private final Map<Integer, ImageGif> gifs = new ConcurrentHashMap<>();
+    private final Map<Integer, ImagemGif> gifs = new ConcurrentHashMap<>();
 
     @Override
     public void inicializar(final Programa programa, final List<Biblioteca> bibliotecasReservadas) throws ErroExecucaoBiblioteca, InterruptedException
@@ -463,7 +455,7 @@ public final class Graficos extends Biblioteca implements Teclado.InstaladorTecl
 
         if (arquivo.exists() && !gifs.containsKey(endereco))
         {
-            gifs.put(endereco, new ImageGif(caminho));
+            gifs.put(endereco, new ImagemGif(arquivo));
         }
         else
         {
@@ -491,7 +483,7 @@ public final class Graficos extends Biblioteca implements Teclado.InstaladorTecl
             throw new ErroExecucaoBiblioteca("O endereço de memória especificado não aponta para uma imagem");
         }
         
-        ImageGif imageGif = gifs.get(endereco);        
+        ImagemGif imageGif = gifs.get(endereco);        
         imageGif.nextImage();
     }
 
@@ -1064,7 +1056,7 @@ public final class Graficos extends Biblioteca implements Teclado.InstaladorTecl
             throw new ErroExecucaoBiblioteca("O endereço de memória especificado não aponta para um gif");
         }
         
-        BufferedImage bf = criarImagemCompativel(gifs.get(endereco).gifFrames.get(quadro).getImage());
+        BufferedImage bf = criarImagemCompativel(gifs.get(endereco).gifFrames.get(quadro).imagem);
         return cacheImagens.adicionarImagem(bf);
     }
     
@@ -1522,215 +1514,5 @@ public final class Graficos extends Biblioteca implements Teclado.InstaladorTecl
     public void sair_modo_tela_cheia() throws ErroExecucaoBiblioteca, InterruptedException
     {
         janela().sairModoTelaCheia();
-    }
-    
-    private class ImageGif
-    {
-        private List<ImageFrame> gifFrames;
-        private int actualImage;
-        public long drawTime = 0;
-
-        public ImageGif(String caminho) throws ErroExecucaoBiblioteca
-        {
-            File arquivo = programa.resolverCaminho(new File(caminho));
-
-            if (arquivo.exists())
-            {
-                gifFrames = new ArrayList<>();
-                try
-                {                   
-                    gifFrames.addAll(readGif(arquivo));
-                }
-                catch (IOException excecao)
-                {
-                    throw new ErroExecucaoBiblioteca(String.format("Ocorreu um erro ao carregar a imagem '%s'", caminho));
-                }
-            }
-            else
-            {
-                throw new ErroExecucaoBiblioteca(String.format("A imagem '%s' não foi encontrada", caminho));
-            }
-        }
-
-        public BufferedImage getActualImage()
-        {
-            return gifFrames.get(actualImage).getImage();
-        }
-        
-        public int getActualNumber()
-        {
-            return actualImage;
-        }
-        
-        public int getGifDelay()
-        {
-            return gifFrames.get(actualImage).getDelay();
-        }
-
-        public void setActualImage(int actualImage)
-        {
-            this.actualImage = actualImage;
-        }
-        
-        public int getFrameNumber()
-        {
-            return gifFrames.size();
-        }
-        
-        public void nextImage()
-        {
-            actualImage = (actualImage+1)%gifFrames.size();
-        }
-        
-        private List<ImageFrame> readGif(File stream) throws IOException{
-            ArrayList<ImageFrame> frames = new ArrayList<ImageFrame>(2);
-
-            ImageReader reader = (ImageReader) ImageIO.getImageReadersByFormatName("gif").next();
-            reader.setInput(ImageIO.createImageInputStream(stream));
-
-            int lastx = 0;
-            int lasty = 0;
-
-            int width = -1;
-            int height = -1;
-
-            IIOMetadata metadata = reader.getStreamMetadata();
-
-            Color backgroundColor = null;
-
-            if(metadata != null) {
-            IIOMetadataNode globalRoot = (IIOMetadataNode) metadata.getAsTree(metadata.getNativeMetadataFormatName());
-
-            NodeList globalColorTable = globalRoot.getElementsByTagName("GlobalColorTable");
-            NodeList globalScreeDescriptor = globalRoot.getElementsByTagName("LogicalScreenDescriptor");
-
-            if (globalScreeDescriptor != null && globalScreeDescriptor.getLength() > 0){
-                IIOMetadataNode screenDescriptor = (IIOMetadataNode) globalScreeDescriptor.item(0);
-
-                if (screenDescriptor != null){
-                    width = Integer.parseInt(screenDescriptor.getAttribute("logicalScreenWidth"));
-                    height = Integer.parseInt(screenDescriptor.getAttribute("logicalScreenHeight"));
-                }
-            }
-
-            if (globalColorTable != null && globalColorTable.getLength() > 0){
-                IIOMetadataNode colorTable = (IIOMetadataNode) globalColorTable.item(0);
-
-                if (colorTable != null) {
-                    String bgIndex = colorTable.getAttribute("backgroundColorIndex");
-
-                    IIOMetadataNode colorEntry = (IIOMetadataNode) colorTable.getFirstChild();
-                    while (colorEntry != null) {
-                        if (colorEntry.getAttribute("index").equals(bgIndex)) {
-                            int red = Integer.parseInt(colorEntry.getAttribute("red"));
-                            int green = Integer.parseInt(colorEntry.getAttribute("green"));
-                            int blue = Integer.parseInt(colorEntry.getAttribute("blue"));
-
-                            backgroundColor = new Color(red, green, blue);
-                            break;
-                        }
-
-                        colorEntry = (IIOMetadataNode) colorEntry.getNextSibling();
-                    }
-                }
-            }
-            }
- 
-            BufferedImage master = null;
-            boolean hasBackround = false;
-            int frameCount = reader.getNumImages(true);
-            
-            for (int frameIndex = 0; frameIndex < frameCount; frameIndex++) {
-                BufferedImage image;
-                try{
-                    image = reader.read(frameIndex);
-                }catch (IndexOutOfBoundsException io){
-                    continue;
-                }
-
-                if (width == -1 || height == -1){
-                    width = image.getWidth();
-                    height = image.getHeight();
-                }
-
-                IIOMetadataNode root = (IIOMetadataNode) reader.getImageMetadata(frameIndex).getAsTree("javax_imageio_gif_image_1.0");
-                IIOMetadataNode gce = (IIOMetadataNode) root.getElementsByTagName("GraphicControlExtension").item(0);
-                NodeList children = root.getChildNodes();
-
-                int delay = Integer.valueOf(gce.getAttribute("delayTime"));
-
-                String disposal = gce.getAttribute("disposalMethod");
-
-                if (master == null){
-                    master = new BufferedImage(width, height, BufferedImage.TYPE_INT_ARGB);
-                    master.createGraphics().setColor(backgroundColor);
-                    master.createGraphics().fillRect(0, 0, master.getWidth(), master.getHeight());
-
-                hasBackround = image.getWidth() == width && image.getHeight() == height;
-
-                    master.createGraphics().drawImage(image, 0, 0, null);
-                }else{
-                    int x = 0;
-                    int y = 0;
-
-                    for (int nodeIndex = 0; nodeIndex < children.getLength(); nodeIndex++){
-                        Node nodeItem = children.item(nodeIndex);
-
-                        if (nodeItem.getNodeName().equals("ImageDescriptor")){
-                            NamedNodeMap map = nodeItem.getAttributes();
-
-                            x = Integer.valueOf(map.getNamedItem("imageLeftPosition").getNodeValue());
-                            y = Integer.valueOf(map.getNamedItem("imageTopPosition").getNodeValue());
-                        }
-                    }
-
-                    if (disposal.equals("restoreToPrevious")){
-                        BufferedImage from = null;
-                        for (int i = frameIndex - 1; i >= 0; i--){
-                            if (!frames.get(i).getDisposal().equals("restoreToPrevious") || frameIndex == 0){
-                                from = frames.get(i).getImage();
-                                break;
-                            }
-                        }
-
-                        {
-                            ColorModel model = from.getColorModel();
-                            boolean alpha = from.isAlphaPremultiplied();
-                            WritableRaster raster = from.copyData(null);
-                            master = new BufferedImage(model, raster, alpha, null);
-                        }
-                    }else if (disposal.equals("restoreToBackgroundColor") && backgroundColor != null){
-                        if (!hasBackround || frameIndex > 1){
-                            master.createGraphics().fillRect(lastx, lasty, frames.get(frameIndex - 1).getWidth(), frames.get(frameIndex - 1).getHeight());
-                        }
-                    }
-                    master.createGraphics().drawImage(image, x, y, null);
-
-                    lastx = x;
-                    lasty = y;
-                }
-
-                try{
-                    BufferedImage copy;
-
-                    {
-                        ColorModel model = master.getColorModel();
-                        boolean alpha = master.isAlphaPremultiplied();
-                        WritableRaster raster = master.copyData(null);
-                        copy = new BufferedImage(model, raster, alpha, null);
-                    }
-                    frames.add(new ImageFrame(copy, delay, disposal, image.getWidth(), image.getHeight()));
-                }
-                catch (Throwable ex)
-                {
-                    ex.printStackTrace(System.err);
-                }
-
-                master.flush();
-            }
-            reader.dispose();
-
-            return frames;
-        }
     }
 }
